backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx
transformers
torch
sentence-transformers
